% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add.missing.genes}
\alias{add.missing.genes}
\title{Add mising genes to the gene expression matrix (for multiple datasets integraiton)}
\usage{
add.missing.genes(epxrM, all.genes)
}
\arguments{
\item{epxrM}{expression matrix in data.frame, matrix or dgCMatrix format}

\item{all.genes}{full list of genes}
}
\value{
A gene expression matrix in dgCMatrix format
}
\description{
Add mising genes to the gene expression matrix (for multiple datasets integraiton)
}
