% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{Add3LevelsAssay}
\alias{Add3LevelsAssay}
\title{Add low, mid, high assays to seuratObj, for comprehensive marker identification}
\usage{
Add3LevelsAssay(seuratObj, slot = "data", assay = "RNA", verbose = F)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{p_value}{The number of resolution for searching}

\item{min_avg_logFC}{Minimal threshold for logFC to pick markers}

\item{min_correlation}{Minimal threshold for correlation to pick markers}
}
\value{
A list. b1 is the positive marker of cluster 0, b2 is the positive marker of cluster 1.
}
\description{
Add low, mid, high assays to seuratObj, for comprehensive marker identification
}
