% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcated_clustering.R
\name{Bt2mBifucation.graph}
\alias{Bt2mBifucation.graph}
\title{Bifurcation based on graph-based clustering}
\usage{
Bt2mBifucation.graph(
  seuratObj,
  cAlgorithm = 4,
  graph.name = "RNA_snn",
  slot = "data",
  assay = "RNA",
  force = T,
  verbose = F
)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{resolution.sets}{The number of resolution for searching}
}
\value{
A bifurcated Seurat object (see active.ident).
}
\description{
Bifurcation based on graph-based clustering
}
