% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterbi_main.R
\name{dataframe_to_vector}
\alias{dataframe_to_vector}
\title{transform dataframe to vector}
\usage{
dataframe_to_vector(df)
}
\arguments{
\item{df}{A dataframe from Seuratobj[["attr"]]}
}
\value{
A vector with names
}
\description{
transform dataframe to vector
}
