% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt2m_main.R
\name{DataframeToVector}
\alias{DataframeToVector}
\title{transform dataframe to vector}
\usage{
DataframeToVector(df)
}
\arguments{
\item{df}{A dataframe from Seuratobj[["attr"]]}
}
\value{
A vector with names
}
\description{
transform dataframe to vector
}
