% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{DEG.cluster.list}
\alias{DEG.cluster.list}
\title{A general function to identify DEGs between case and control}
\usage{
DEG.cluster.list(
  seuratObj,
  cluster.list,
  ident.1 = "Vcl cKO",
  ident.2 = "Control",
  assay = "RNA"
)
}
\description{
A general function to identify DEGs between case and control
}
