% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{discreteOrContinuous}
\alias{discreteOrContinuous}
\title{judge the discrete and continuous cell 2}
\usage{
discreteOrContinuous(seuratObj)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{slot}{The method to perform bifurcation clustering "graph (default), hclust or kmeans"}

\item{assay}{Minimal number of markers to confirm a bifurcation}
}
\value{
A list. cellMeta contains the preliminary bifurcation for each level
marker_chain contains all the significant markers for each cluster
bifucation contains the bifurcation details (parent, child1, child2)
}
\description{
judge the discrete and continuous cell 2
}
