% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{draw_GO_barplot}
\alias{draw_GO_barplot}
\title{Draw barplot for GO annotation}
\usage{
draw_GO_barplot(barplot_df, cluster.chain)
}
\arguments{
\item{barplot_df}{A GO annotation dataframe from clusterProfiler}

\item{cluster.chain}{cluster.chain from iterbi}
}
\value{
A ggplot barplot
}
\description{
Draw barplot for GO annotation
}
