% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{draw_iterbi_cluster_tree}
\alias{draw_iterbi_cluster_tree}
\title{Draw cluster tree/chain by clustree}
\usage{
draw_iterbi_cluster_tree(
  seuratObj,
  iterbi.cellMeta,
  node_text_size = 7,
  node_size = 8
)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{iterbi.cellMeta}{iterbi.cellMeta from iterbi}

\item{node_text_size}{see node_text_size in clustree function}

\item{node_size}{see node_size in clustree function}
}
\value{
clustree object
}
\description{
Draw cluster tree/chain by clustree
}
