% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{draw_marker_chain_dotplot}
\alias{draw_marker_chain_dotplot}
\title{Draw cluster tree/chain by clustree}
\usage{
draw_marker_chain_dotplot(
  seuratObj,
  iterbi.marker.chain,
  rmDup = T,
  top_n = 3,
  rel_heights = c(0.5, 9, 1.5),
  rel_widths = c(1, 4)
)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{iterbi.marker.chain}{iterbi.marker.chain from iterbi}

\item{rmDup}{remove duplicated genes or not}

\item{top_n}{top n genes for dotplot (sort by P-value)}

\item{rel_heights}{relative heights for plot_grid(), the first and last control the height of top and bottom margin, the middle one controls the height of tree}

\item{rel_widths}{relative widths for plot_grid(), the first one controls the widths of tree, the second one controls the widths of dotplot}
}
\value{
A plot_grid integrated tree and dotplot
}
\description{
Draw cluster tree/chain by clustree
}
