% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{DrawBarplotGO}
\alias{DrawBarplotGO}
\title{Draw barplot for GO annotation}
\usage{
DrawBarplotGO(barplot_df, cluster.chain)
}
\arguments{
\item{barplot_df}{A GO annotation dataframe from clusterProfiler}

\item{cluster.chain}{cluster.chain from bt2m}
}
\value{
A ggplot barplot
}
\description{
Draw barplot for GO annotation
}
