% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{DrawGOchain}
\alias{DrawGOchain}
\title{Get GO chain corresponding to a cluster chain}
\usage{
DrawGOchain(bt2m.GO.anno, cluster.chain, top_n = 3)
}
\arguments{
\item{bt2m.GO.anno}{A GO annotation dataframe from clusterProfiler}

\item{cluster.chain}{cluster.chain from bt2m}

\item{top_n}{top n GO terms for barplot (sort by P-value)}
}
\value{
A ggplot barplot showed GO chain
}
\description{
Get GO chain corresponding to a cluster chain
}
