% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{EstimateDropoutRate}
\alias{EstimateDropoutRate}
\title{Estimate dropout rate of single-cell RNA-seq data by house keeping genes}
\usage{
EstimateDropoutRate(seuratObj, top = 15, hkg)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{p_value}{The number of resolution for searching}

\item{min_avg_logFC}{Minimal threshold for logFC to pick markers}

\item{min_correlation}{Minimal threshold for correlation to pick markers}
}
\value{
A list. b1 is the positive marker of cluster 0, b2 is the positive marker of cluster 1.
}
\description{
Estimate dropout rate of single-cell RNA-seq data by house keeping genes
}
