% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcated_clustering.R
\name{FindBifurcationResolution}
\alias{FindBifurcationResolution}
\title{Find the best resolution for bifurcation in graph-based clustering}
\usage{
FindBifurcationResolution(seuratObj, resolution.sets = 50)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{resolution.sets}{The number of resolution for searching}
}
\value{
The best resolution which can bifurcate all cells
}
\description{
Find the best resolution for bifurcation in graph-based clustering
}
