% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{FindBinaryMarkers}
\alias{FindBinaryMarkers}
\title{Identify binary markers}
\usage{
FindBinaryMarkers(
  seuratObj,
  method = "presto",
  inputFactor = F,
  slot = "data",
  assay = "RNA",
  p_value = 0.05,
  min_avg_log2FC = 0.1,
  min_correlation = 0.2,
  verbose = F
)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{p_value}{The number of resolution for searching}

\item{min_correlation}{Minimal threshold for correlation to pick markers}

\item{min_avg_logFC}{Minimal threshold for logFC to pick markers}
}
\value{
A list. b1 is the positive marker of cluster 0, b2 is the positive marker of cluster 1.
}
\description{
Identify binary markers
}
