% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene.anno.GRCh38.v3}
\alias{gene.anno.GRCh38.v3}
\title{Gene annotation for human genome}
\format{
A vector
}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/software/downloads/latest}
}
\usage{
gene.anno.GRCh38.v3
}
\description{
All human genes and annotation
}
\concept{data}
\keyword{datasets}
