% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_marker_GO_chain}
\alias{get_marker_GO_chain}
\title{Get GO chain corresponding to a cluster chain}
\usage{
get_marker_GO_chain(iterbi.GO.anno, cluster.chain, top_n = 3)
}
\arguments{
\item{iterbi.GO.anno}{A GO annotation dataframe from clusterProfiler}

\item{cluster.chain}{cluster.chain from iterbi}

\item{top_n}{top n GO terms for barplot (sort by P-value)}
}
\value{
A ggplot barplot showed GO chain
}
\description{
Get GO chain corresponding to a cluster chain
}
