% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_test_data}
\alias{get_test_data}
\title{Download raw single-cell matrix for testing}
\usage{
get_test_data()
}
\value{
dowloaded single-cell dataset
}
\description{
Download raw single-cell matrix for testing
}
