% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterbi_main.R
\name{get_unique_marker}
\alias{get_unique_marker}
\title{filter uniquely expressed markers by expression percentage}
\usage{
get_unique_marker(
  seuratObj,
  iterbi.marker.chain,
  assay = "RNA",
  slot = "data",
  min_cluster_pct = 0.3,
  max_bcg_pct = 0.1,
  min_diff = 0.3
)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{iterbi.marker.chain}{iterbi.marker.chain dataframe contains all the markers}

\item{assay}{Assay used for prediction}

\item{slot}{slot used for prediction}

\item{min_cluster_pct}{Minimal expression percentage of target cluster}

\item{max_bcg_pct}{Maximum expression percentage of the background cells (non-target cells)}

\item{min_diff}{Minimal difference (expression percentage) between target cluster and background cells}
}
\value{
Uniquely expressed iterbi.marker.chain
}
\description{
filter uniquely expressed markers by expression percentage
}
