% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterbi_main.R
\name{GetClusterChain}
\alias{GetClusterChain}
\title{Find the cluster chain of any given cluster}
\usage{
GetClusterChain(iterbi.cellMeta, iterbi.bifucation, target.cluster)
}
\arguments{
\item{iterbi.cellMeta}{iterbi.cellMeta from iterbi}

\item{iterbi.bifucation}{iterbi.bifucation from iterbi}

\item{target.cluster}{target cluster}
}
\value{
A vector, names are cluster name, element shows the states of clusters
}
\description{
Find the cluster chain of any given cluster
}
