% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{GetClusterChainNoName}
\alias{GetClusterChainNoName}
\title{Find the cluster chain of any given cluster, direct output cluster chain}
\usage{
GetClusterChainNoName(bt2m.cellMeta, target.cluster)
}
\arguments{
\item{bt2m.cellMeta}{bt2m.cellMeta from bt2m}

\item{target.cluster}{target cluster}

\item{bt2m.bifucation}{bt2m.bifucation from bt2m}
}
\value{
A vector, names are cluster name, element shows the states of clusters
}
\description{
Find the cluster chain of any given cluster, direct output cluster chain
}
