% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{GetTestData}
\alias{GetTestData}
\title{Download raw single-cell matrix for testing}
\usage{
GetTestData()
}
\value{
dowloaded single-cell dataset
}
\description{
Download raw single-cell matrix for testing
}
