% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterbi_main.R
\name{GetUniqueMarker}
\alias{GetUniqueMarker}
\title{filter uniquely expressed markers by expression percentage}
\usage{
GetUniqueMarker(
  seuratObj,
  iterbi.marker.chain,
  assay = "RNA",
  slot = "data",
  min.cluster.pct = 0.3,
  max.bcg.pct = 0.1,
  min.diff = 0.3
)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{iterbi.marker.chain}{iterbi.marker.chain dataframe contains all the markers}

\item{assay}{Assay used for prediction}

\item{slot}{slot used for prediction}

\item{min.cluster.pct}{Minimal expression percentage of target cluster}

\item{max.bcg.pct}{Maximum expression percentage of the background cells (non-target cells)}

\item{min.diff}{Minimal difference (expression percentage) between target cluster and background cells}
}
\value{
Uniquely expressed iterbi.marker.chain
}
\description{
filter uniquely expressed markers by expression percentage
}
