% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HGNC.gene.pairs}
\alias{HGNC.gene.pairs}
\title{Mouse transcription factors}
\format{
A vector
}
\source{
\url{http://bioinfo.life.hust.edu.cn/AnimalTFDB/#!/download}
}
\usage{
HGNC.gene.pairs
}
\description{
A vector of genes used in marker annotation
}
\concept{data}
\keyword{datasets}
