% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human.house.keeping.gene}
\alias{human.house.keeping.gene}
\title{Human house keeping genes}
\format{
A list of two vectors
\describe{
  \item{mouse.ligand}{ligand genes in mouse}
  \item{mouse.receptor}{receptor genes in mouse}
}
}
\source{
\url{http://www.cellchat.org/}
}
\usage{
human.house.keeping.gene
}
\description{
A list of genes used in marker annotation
}
\concept{data}
\keyword{datasets}
