% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{identify_binary_markers}
\alias{identify_binary_markers}
\title{Identify binary markers}
\usage{
identify_binary_markers(
  seuratObj,
  p_value = 0.01,
  min_avg_logFC = 0.1,
  min_correlation = 0.3
)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{p_value}{The number of resolution for searching}

\item{min_avg_logFC}{Minimal threshold for logFC to pick markers}

\item{min_correlation}{Minimal threshold for correlation to pick markers}
}
\value{
A list. b1 is the positive marker of cluster 0, b2 is the positive marker of cluster 1.
}
\description{
Identify binary markers
}
