% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{InitializeFeatureAnno}
\alias{InitializeFeatureAnno}
\title{Intialize feature annotation data for each assay}
\usage{
InitializeFeatureAnno(seuratObj, slot = "count", assay = "RNA")
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{slot}{The method to perform bifurcation clustering "graph (default), hclust or kmeans"}

\item{assay}{Minimal number of markers to confirm a bifurcation}
}
\value{
A list. cellMeta contains the preliminary bifurcation for each level
marker_chain contains all the significant markers for each cluster
bifucation contains the bifurcation details (parent, child1, child2)
}
\description{
Intialize feature annotation data for each assay
}
