% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcated_clustering.R
\name{iterbi.bifucation.graph}
\alias{iterbi.bifucation.graph}
\title{Bifurcation based on graph-based clustering}
\usage{
iterbi.bifucation.graph(seuratObj, res_sets = 50)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{res_sets}{The number of resolution for searching}
}
\value{
A bifurcated Seurat object (see active.ident).
}
\description{
Bifurcation based on graph-based clustering
}
