% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcated_clustering.R
\name{iterbi.bifucation.kmeans}
\alias{iterbi.bifucation.kmeans}
\title{Bifurcation based on K-means clustering}
\usage{
iterbi.bifucation.kmeans(seuratObj, method = "euclidean")
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{method}{The distance measurement method "euclidean or correlation"}
}
\value{
A bifurcated Seurat object (see active.ident).
}
\description{
Bifurcation based on K-means clustering
}
