% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcated_clustering.R
\name{IterbiBifucation.graph}
\alias{IterbiBifucation.graph}
\title{Bifurcation based on graph-based clustering}
\usage{
IterbiBifucation.graph(seuratObj, resolution.sets = 50)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{resolution.sets}{The number of resolution for searching}
}
\value{
A bifurcated Seurat object (see active.ident).
}
\description{
Bifurcation based on graph-based clustering
}
