% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcated_clustering.R
\name{IterbiBifucation.hclust}
\alias{IterbiBifucation.hclust}
\title{Bifurcation based on hierarchical clustering}
\usage{
IterbiBifucation.hclust(seuratObj, method = "euclidean")
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{method}{The distance measurement method "euclidean or correlation"}
}
\value{
A bifurcated Seurat object (see active.ident).
}
\description{
Bifurcation based on hierarchical clustering
}
