% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterbi_main.R
\name{IterbiEnrichGO}
\alias{IterbiEnrichGO}
\title{Perform GO annotation of iterbi.marker.chain dataframe}
\usage{
IterbiEnrichGO(
  iterbi.marker.chain,
  organism = "hs",
  pvalueCutoff = 0.05,
  min_count = 5
)
}
\arguments{
\item{iterbi.marker.chain}{terbi.marker.chain dataframe contains all the markers}

\item{organism}{"hs" for Homo sapiens, or "mm" for Mus musculus}

\item{pvalueCutoff}{cut off P-value for GO annotations}

\item{min_count}{Minimal count of genes for GO annotations}
}
\value{
GO annotation dataframe labeled with cluster
}
\description{
Perform GO annotation of iterbi.marker.chain dataframe
}
