% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{JASPAR.motif.rmDup}
\alias{JASPAR.motif.rmDup}
\title{Human ligand and receptor}
\format{
A list of two vectors
\describe{
  \item{human.ligand}{ligand genes in human}
  \item{human.receptor}{receptor genes in human}
}
}
\source{
\url{http://www.cellchat.org/}
}
\usage{
JASPAR.motif.rmDup
}
\description{
A list of genes used in marker annotation
}
\concept{data}
\keyword{datasets}
