% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lr.pair.all.uniq.hs}
\alias{lr.pair.all.uniq.hs}
\title{Mouse ligand and receptor}
\format{
A list of two vectors
\describe{
  \item{mouse.ligand}{ligand genes in mouse}
  \item{mouse.receptor}{receptor genes in mouse}
}
}
\source{
\url{http://www.cellchat.org/}
}
\usage{
lr.pair.all.uniq.hs
}
\description{
A list of genes used in marker annotation
}
\concept{data}
\keyword{datasets}
