% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m6A.regulators}
\alias{m6A.regulators}
\title{m6A writer, eraser and reader in human and mouse}
\format{
A vector
}
\source{
\url{https://housekeeping.unicamp.br/?download}
}
\usage{
m6A.regulators
}
\description{
m6A.regulators
}
\concept{data}
\keyword{datasets}
