% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MGI.gene.pairs}
\alias{MGI.gene.pairs}
\title{Human transcription factors}
\format{
A vector
}
\source{
\url{http://bioinfo.life.hust.edu.cn/HumanTFDB/#!/download}
}
\usage{
MGI.gene.pairs
}
\description{
A vector of genes used in marker annotation
}
\concept{data}
\keyword{datasets}
