% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterbi_main.R
\name{ora.go.kegg.clusterProfiler}
\alias{ora.go.kegg.clusterProfiler}
\title{GO KEGG ORA analysis by clusterProfiler}
\usage{
ora.go.kegg.clusterProfiler(geneList = markerList, organism = "hs")
}
\arguments{
\item{geneList}{a gene list}

\item{organism}{"hs" for Homo sapiens, or "mm" for Mus musculus}
}
\value{
a list with GO and KEGG annotation result (clusterProfiler format)
}
\description{
GO KEGG ORA analysis by clusterProfiler
}
