% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PrepareExpressionMatrix}
\alias{PrepareExpressionMatrix}
\title{Prepare expression matrix for heatmap visualization, estimate the overall expression of marker modules}
\usage{
PrepareExpressionMatrix(
  seuratObj,
  iterbi.marker.chain,
  assay = "RNA",
  slot = "scale.data",
  known_markers = c()
)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{iterbi.marker.chain}{iterbi.marker.chain from iterbi}

\item{assay}{Assay used for prediction}

\item{slot}{slot used for prediction}

\item{known_markers}{the genes to include in the heatmap (like known markers)}
}
\value{
A expression matrix contain the average expression of marker modules and known markers
}
\description{
Prepare expression matrix for heatmap visualization, estimate the overall expression of marker modules
}
