% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterbi_main.R
\name{remove_duplicated_GO}
\alias{remove_duplicated_GO}
\title{remove duplicated GO terms based on overlaps of genes}
\usage{
remove_duplicated_GO(tmp.GO.df, max.overlap = 0.6)
}
\arguments{
\item{tmp.GO.df}{GO annotation dataframe from clusterProfiler package}

\item{max.overlap}{Maximum overlapped percentage of genes}
}
\value{
GO annotation dataframe without duplications
}
\description{
remove duplicated GO terms based on overlaps of genes
}
