% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{RemoveDuplicatedMarker}
\alias{RemoveDuplicatedMarker}
\title{Remove duplicated markers from last level}
\usage{
RemoveDuplicatedMarker(iterbi.marker.chain)
}
\arguments{
\item{iterbi.marker.chain}{iterbi.marker.chain from iterbi}
}
\value{
iterbi.marker.chain without duplicated genes
}
\description{
Remove duplicated markers from last level
}
