% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterbi_main.R
\name{rename.iterbi}
\alias{rename.iterbi}
\title{rename the clusters inside the iterbi result}
\usage{
rename.iterbi(iterbi.result)
}
\arguments{
\item{iterbi.result}{A result file from run.iterbi() function}
}
\value{
A renamed list. cellMeta contains the final bifurcation for each level
marker_chain contains all the significant markers for each cluster
bifucation contains the bifurcation details (parent, child1, child2)
}
\description{
rename the clusters inside the iterbi result
}
