% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{RenameBT2M}
\alias{RenameBT2M}
\title{rename the clusters inside the bt2m result}
\usage{
RenameBT2M(bt2m.result)
}
\arguments{
\item{bt2m.result}{A result file from RunBt2m() function}
}
\value{
A renamed list. cellMeta contains the final bifurcation for each level
marker_chain contains all the significant markers for each cluster
bifucation contains the bifurcation details (parent, child1, child2)
}
\description{
rename the clusters inside the bt2m result
}
