% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{RunBT2M.AC.supervise}
\alias{RunBT2M.AC.supervise}
\title{Start from known cell type annotation RunBT2M.AC.supervise}
\usage{
RunBT2M.AC.supervise(seuratObj, group, reduction = "umap", verbose = F)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{verbose}{Print detail proccessing messages}

\item{method}{The method to perform bifurcation clustering "graph (default), hclust or kmeans"}

\item{min.marker.num}{Minimal number of markers to confirm a bifurcation}

\item{max.level.num}{Maximum number of level for bifurcation}

\item{min.cell.count}{Minimal number of cells to perform bifurcation (must bigger than PC number: 50)}

\item{resolution.sets}{The number of resolution for searching}
}
\value{
A list. cellMeta contains the preliminary bifurcation for each level
marker_chain contains all the significant markers for each cluster
bifucation contains the bifurcation details (parent, child1, child2)
}
\description{
Start from known cell type annotation RunBT2M.AC.supervise
}
