% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{subset_cells}
\alias{subset_cells}
\title{a subset function for seurat obj, origninal cannot work sometimes}
\usage{
subset_cells(seuratObj, condition)
}
\description{
a subset function for seurat obj, origninal cannot work sometimes
}
