% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{subsetSeuratObjByCells}
\alias{subsetSeuratObjByCells}
\title{Subset seurat object by cell names}
\usage{
subsetSeuratObjByCells(seuratObj, tmp.cells)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{tmp.cells}{cells for subseting}
}
\value{
A subset of seurat object
}
\description{
Subset seurat object by cell names
}
