% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{write_iterbi_to_seurat}
\alias{write_iterbi_to_seurat}
\title{Write iterbi result to Seurat object}
\usage{
write_iterbi_to_seurat(seuratObj, iterbi.result)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{iterbi.result}{iterbi.result}
}
\value{
A Seurat object contains iterbi.result. iterbi will be stored in assay data region of Seurat object ("seuratObj@assays$iterbi")
}
\description{
Write iterbi result to Seurat object
}
