% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterbi_main.R
\name{AddMarkerExpressionPct}
\alias{AddMarkerExpressionPct}
\title{filter uniquely expressed markers by expression percentage}
\usage{
AddMarkerExpressionPct(
  seuratObj,
  iterbi.marker.chain,
  assay = "RNA",
  slot = "data"
)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{iterbi.marker.chain}{iterbi.marker.chain dataframe contains all the markers}

\item{assay}{Assay used for prediction}

\item{slot}{slot used for prediction}
}
\value{
A new dataframe with two additional columns: cluster_pct (expression percentage in the cluster) and bcg_pct (expression percentage in the background cells)
}
\description{
filter uniquely expressed markers by expression percentage
}
