% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{DrawMarkerChainHeatmap}
\alias{DrawMarkerChainHeatmap}
\title{Draw heatmap by ComplexHeatmap}
\usage{
DrawMarkerChainHeatmap(
  seuratObj,
  bt2m.cellMeta,
  bt2m.marker.chain,
  compare_anno = "",
  known_markers = c()
)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{bt2m.cellMeta}{bt2m.cellMeta from bt2m}

\item{bt2m.marker.chain}{bt2m.marker.chain from bt2m}

\item{compare_anno}{select a annotation in seurat object for comparision (e.g. previous annotation)}

\item{known_markers}{the genes to include in the heatmap (like known markers)}
}
\value{
ComplexHeatmap object
}
\description{
Draw heatmap by ComplexHeatmap
}
