% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt2m_main.R
\name{GetClusterChain}
\alias{GetClusterChain}
\title{Find the cluster chain of any given cluster}
\usage{
GetClusterChain(bt2m.cellMeta, bt2m.bifucation, target.cluster)
}
\arguments{
\item{bt2m.cellMeta}{bt2m.cellMeta from bt2m}

\item{bt2m.bifucation}{bt2m.bifucation from bt2m}

\item{target.cluster}{target cluster}
}
\value{
A vector, names are cluster name, element shows the states of clusters
}
\description{
Find the cluster chain of any given cluster
}
