% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterbi_main.R
\name{GetInitialBifurcation}
\alias{GetInitialBifurcation}
\title{Find initial bifurcation event of any two clusters in iterbi.cellMeta}
\usage{
GetInitialBifurcation(iterbi.cellMeta, cluster_1, cluster_2, balance_cells = T)
}
\arguments{
\item{iterbi.cellMeta}{iterbi.cellMeta from iterbi}

\item{cluster_1}{cluster 1}

\item{cluster_2}{cluster 2}

\item{balance_cells}{output balanced cell number (T or F)}
}
\value{
A list. subset_seuratObj is the subset cells in cluster 1 and cluster 2
diff_marker_1 is the marker of cluster 1, diff_marker_2 is the marker of cluster 2
}
\description{
Find initial bifurcation event of any two clusters in iterbi.cellMeta
}
