% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterbi_main.R
\name{RemoveDuplicatedMarker}
\alias{RemoveDuplicatedMarker}
\title{Remove duplicated markers (just for plotting)}
\usage{
RemoveDuplicatedMarker(iterbi.marker.chain, method = "correlation")
}
\arguments{
\item{iterbi.marker.chain}{iterbi.marker.chain from iterbi}

\item{method}{select a method to sort markers}
}
\value{
iterbi.marker.chain without duplicated genes
}
\description{
Remove duplicated markers (just for plotting)
}
