% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterbi_main.R
\name{RemoveDuplicatedMarker}
\alias{RemoveDuplicatedMarker}
\title{Remove duplicated markers (just for plotting)}
\usage{
RemoveDuplicatedMarker(bt2m.marker.chain, method = "correlation")
}
\arguments{
\item{bt2m.marker.chain}{bt2m.marker.chain from bt2m}

\item{method}{select a method to sort markers}
}
\value{
bt2m.marker.chain without duplicated genes
}
\description{
Remove duplicated markers (just for plotting)
}
