% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterbi_main.R
\name{WriteIterbiIntoSeurat}
\alias{WriteIterbiIntoSeurat}
\title{Write bt2m result to Seurat object}
\usage{
WriteIterbiIntoSeurat(seuratObj, bt2m.result)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{bt2m.result}{bt2m.result}
}
\value{
A Seurat object contains bt2m.result. bt2m will be stored in assay data region of Seurat object ("seuratObj@assays$bt2m")
}
\description{
Write bt2m result to Seurat object
}
